+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 # CHANGELOG
 
+## 0.8.0 - 2025-06-01
+
+* Adjusted some method modifiers and added return types
+* Fixed signature generation with duplicate query parameters
+
 ## 0.7.0 - 2025-06-01
 
 * Dropped support for HHVM and PHP <7.2.5
+* Dropped support for Guzzle 6.x and PSR-7 1.x
 * Added support for PHP 8.1, 8.2, 8.3, 8.4
+* Add param types to various methods
 
 ## 0.6.0 - 2021-07-13